requirements.txt
aiogram==3.0.0b7
numpy~=1.24.0
Pillow~=10.0.0
openai